REGRESSION (r239306): Don't disable font smoothing in transparent layers on macOS Mojave and later https://bugs.webkit.org/show_bug.cgi?id=193095 <rdar://problem/47014944>
Reviewed by Zalan Bujtas.

Source/WebCore:

In r239306 we stopped making CALayers for font-smoothed text when not necessary on macOS Mojave
and later. However, we still turned off smoothing for non-opaque layers (setShouldSmoothFonts(false)),
which continues to affect the appearance of text.

Fix by only calling setShouldSmoothFonts(false) when the OS supports font smoothing.

Test: compositing/contents-format/subpixel-antialiased-text.html

* platform/graphics/ca/cocoa/PlatformCALayerCocoa.mm:
(WebCore::PlatformCALayer::drawLayerContents):

LayoutTests:

A ref test that compares the rendering of layer and non-layer text with smoothing and antialiasing.

* compositing/contents-format/subpixel-antialiased-text-expected.html: Added.
* compositing/contents-format/subpixel-antialiased-text.html: Added.
* platform/mac/TestExpectations:

Identifier: 207621@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239591 268f45cc-cd09-0410-ab3c-d52691b4dbfc