REGRESSION (r238090): After showing the Find banner or tab bar, hit tests are vertically offset (or painting is not offset) https://bugs.webkit.org/show_bug.cgi?id=193124 rdar://problem/46755409
Reviewed by Tim Horton.

Top content inset feeds into scrolling tree geometry, so when it changes we need to trigger
an update of the root scrolling node, which happens via RenderLayerBacking::updateGeometry().
So set a dirty bit on the root layer in frameViewDidChangeSize(), which is called from the code
path that runs when top content inset changes.

Find banner behavior is not easily testable. platform/mac/fast/events/content-inset-hit-testing.html did not detect the regression.

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::frameViewDidChangeSize):


Identifier: 207629@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239601 268f45cc-cd09-0410-ab3c-d52691b4dbfc