REGRESSION (r238522): Erratic scrolling on Google flights search result page and vrbo.com https://bugs.webkit.org/show_bug.cgi?id=192996 <rdar://problem/46573552>
Reviewed by Simon Fraser.

Source/WebCore:

Only scroll a text field if its inner text size changes and it is the currently active and
focused element on the page.

Test: fast/scrolling/page-should-not-scroll-on-unfocused-text-field-layout.html

* rendering/RenderTextControlSingleLine.cpp:
(WebCore::RenderTextControlSingleLine::layout):

LayoutTests:

Add a test to ensure that we do not scroll the page when we lay out an unfocused text field.

* fast/scrolling/page-should-not-scroll-on-unfocused-text-field-layout-expected.txt: Added.
* fast/scrolling/page-should-not-scroll-on-unfocused-text-field-layout.html: Added.

Identifier: 207680@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239652 268f45cc-cd09-0410-ab3c-d52691b4dbfc