Use unorm2_normalize instead of precomposedStringWithCanonicalMapping in userVisibleString https://bugs.webkit.org/show_bug.cgi?id=192945
Reviewed by Alex Christensen.

Replace use of the nice NSString function precomposedStringWithCanonicalMapping with the ICU
API unorm2_normalize. This is to prep the code for translation to cross-platform C++. Of
course this is much worse than the preexisting code, but this is just a transitional
measure and not the final state of the code. It wouldn't make sense to do this if the code
were to remain Objective C++.

* wtf/cocoa/NSURLExtras.mm:
(WTF::toNormalizationFormC):
(WTF::userVisibleString):

Identifier: 207939@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239970 268f45cc-cd09-0410-ab3c-d52691b4dbfc