[WebAudio] Call AudioContext::uninitialize() immediately when the AudioContext is stopped https://bugs.webkit.org/show_bug.cgi?id=192586
Patch by Yacine Bandou <yacine.bandou@softathome.com> on 2019-01-18
Reviewed by Philippe Normand.

When WebProcess is killed, AudioContext::uninitialize() is not called immediately in the stop so
the AudioDestinationNode is not destroyed.

In my case, I have a resource device manager, the output audio device is reserved when AudioDestinationNode
is instantiated and it is released when AudioDestinationNode is destroyed, thus when the webprocess is killed,
the resources leak.

AudioContext::uninitialize() is not called immediately since r94608.
This modification can now be reverted without regression in WebAudio tests.

Test: webaudio/mediaelementaudiosourcenode-gc.html

* Modules/webaudio/AudioContext.cpp:
(WebCore::AudioContext::stop):

Identifier: 208068@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@240142 268f45cc-cd09-0410-ab3c-d52691b4dbfc