[WebAudio] Release the AudioDestination when uninitializing DefaultAudioDestinationNode https://bugs.webkit.org/show_bug.cgi?id=192590
Patch by Yacine Bandou <yacine.bandou@softathome.com> on 2019-01-18
Reviewed by Philippe Normand.

When we uninitialize DefaultAudioDestinationNode, the AudioDestination is stopped but not destroyed.

On some platforms the resources are allocated and released with the AudioDestination, thus when we uninitialize
DefaultAudioDestinationNode we don't release resources because the AudioDestination is not destroyed.

* Modules/webaudio/DefaultAudioDestinationNode.cpp:
(WebCore::DefaultAudioDestinationNode::uninitialize):

Identifier: 208069@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@240143 268f45cc-cd09-0410-ab3c-d52691b4dbfc