FloatWithRect has invalid and inaccessible default constructor https://bugs.webkit.org/show_bug.cgi?id=194057 <rdar://problem/47685211>
Reviewed by Zalan Bujtas.

FloatWithRect has a default constructor declared as 'default'. This
constructor is invalid because FloatWithRect has a data member that's
a reference that will not get initialized with this constructor. Since
it's invalid, and since it's private and not accessed by anything,
remove this constructor.

No new tests since there should be no observable behavior difference.

* rendering/line/LineLayoutState.h:


Identifier: 208532@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@240744 268f45cc-cd09-0410-ab3c-d52691b4dbfc