[GLIB] WebDriver: race condition when session starts https://bugs.webkit.org/show_bug.cgi?id=194086
Reviewed by Philippe Normand.

This is causing flakiness in the bots, I can only reproduce it locally when my CPUs are at 100%. The thing
is that sometimes we receive an empty target list right after the connection is established because
RemoteInspector::setClient() calls pushListingsSoon() before a target has been registered. Most of the times the
target is registered before the listing is pushed, but it can happen that when pushed, the listing is still
empty. After the target is registered an empty target list means the connection was lost, so we clear the
connection and target. We need to ensure we do that only after the target has been registered.

* glib/SessionHostGlib.cpp:

Identifier: 208559@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@240778 268f45cc-cd09-0410-ab3c-d52691b4dbfc